Split up components into different files.
